requirements.txt
Django
djangorestframework
djangorestframework-jwt
mysqlclient
django-mysql
django-cors-headers
pika
